refactor: extract memory data-access helpers into queries.py
Move get_memory_entries and fetch_memory_content out of index_builder.py
into a dedicated queries module. index_builder now only handles index
formatting, and injector imports directly from queries.
